+++ lib/squeel/adapters/active_record/4.0/relation_extensions.rb
@@ -160,10 +160,12 @@ module Squeel
 
           # if a symbol is given we prepend the quoted table name
           args = args.map { |arg|
-            arg.is_a?(Symbol) ? "#{quoted_table_name}.#{arg} ASC" : arg
+            arg.is_a?(Symbol) ?
+              Arel::Nodes::Ascending.new(klass.arel_table[arg]) :
+              arg
           }
 
-          self.order_values = args + self.order_values
+          self.order_values += args
           self
         end